Refactor services
Refactored service level, now everything works via interfaces.
Also changed Client.passport to long

+++ src/main/java/com/example/sanatorium/models/Client.java
@@ -33,7 +33,7 @@ public class Client {
     private int age;
 
     @Column(name = "passport",nullable = false, unique = true)
-    private int passport;
+    private long passport;
 
     @OneToMany(cascade = CascadeType.ALL, mappedBy = "client")
     private Set<ResidentialApartmentBooking> residentialApartmentBookings;
